chore: cleanly archive dynamic files at build time
