+++ tests/basic/backend.json
@@ -1,15 +1,5 @@
 {
 	"registries": {
-		"default": {
-			"git": {
-				"url": {
-					"https": "https://github.com/rivet-gg/opengb-registry.git",
-					"ssh": "git@github.com:rivet-gg/opengb-registry.git"
-				},
-				"rev": "3c40379899162d848cea88afc58b4dc6ab4cc546",
-				"directory": "./modules"
-			}
-		},
 		"local": {
 			"local": {
 				"directory": "./modules"
